+++ components/collector/src/source_collectors/trivy/security_warnings.py
@@ -29,17 +29,19 @@ class TrivyJSONSecurityWarnings(SecurityWarningsSourceCollector, JSONFileSourceC
             for vulnerability in result.get("Vulnerabilities") or []:
                 vulnerability_id = vulnerability["VulnerabilityID"]
                 package_name = vulnerability["PkgName"]
+                installed_version = vulnerability["InstalledVersion"]
                 references = vulnerability.get("References", [])
                 url = references[0] if references else ""  # Assume the 1st link is at least as relevant as the others
                 entities.append(
                     Entity(
-                        key=f"{vulnerability_id}@{package_name}@{target}",
+                        key=f"{vulnerability_id}@{package_name}@{installed_version}",
+                        old_key=f"{vulnerability_id}@{package_name}@{target}",  # Key changed after v5.50.0 due to https://github.com/ICTU/quality-time/issues/12746
                         vulnerability_id=vulnerability_id,
                         title=vulnerability.get("Title", vulnerability_id),
                         description=vulnerability.get("Description", ""),
                         level=vulnerability["Severity"],
                         package_name=package_name,
-                        installed_version=vulnerability["InstalledVersion"],
+                        installed_version=installed_version,
                         fixed_version=vulnerability.get("FixedVersion", ""),
                         url=url,
                     ),